fix cellcycleanalysis as part of module, several bug fixes to make program more robust, checked with several analysis runs.
