--- 221102_CellCycleProfile_Exp5_inhibitors_RPE1cdk1as/Plate_layout.csv
@@ -1,3 +1,0 @@
-,Well,Well_ID,Cell_Line,Condition
-0,A1,10684,RPE-1,NT
-1,A2,10707,RPE-1,DMSO
