+++ bin/omero_screen_run
